Separate component for Phone Number.
